+++ cmd/logs.go
@@ -41,12 +41,8 @@ func runLogs(cmd *cobra.Command, args []string) error {
 	if flagProject != "" {
 		stack.Project = flagProject
 	}
-	runtimeName := stack.Runtime
-	if flagRuntime != "" {
-		runtimeName = flagRuntime
-	}
 
-	runtime, err := newRuntime(runtimeName)
+	runtime, err := runtimeFrom(stack)
 	if err != nil {
 		return err
 	}